+++ frontend/src/components/prompts/Delete.vue
@@ -11,13 +11,13 @@
       </div>
     </div>
     <div class="card-action">
-      <button @click="closeHovers" class="button button--flat button--grey" :aria-label="$t('buttons.cancel')"
-        :title="$t('buttons.cancel')">
-        {{ $t("buttons.cancel") }}
+      <button @click="closeHovers" class="button button--flat button--grey" :aria-label="$t('general.cancel')"
+        :title="$t('general.cancel')">
+        {{ $t("general.cancel") }}
       </button>
       <button @click="submit" class="button button--flat button--red" aria-label="Confirm-Delete"
-        :title="$t('buttons.delete')">
-        {{ $t("buttons.delete") }}
+        :title="$t('general.delete')">
+        {{ $t("general.delete") }}
       </button>
     </div>
 </template>