+++ frontend/src/components/prompts/MoveCopy.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="card-title">
-    <h2>{{ $t(`prompts.${operation}`) }}</h2>
+    <h2>{{ operation === 'move' ? $t('prompts.move') : $t('prompts.copy') }}</h2>
   </div>
 
   <div class="card-content">
@@ -16,19 +16,20 @@
   </div>
   <div v-if="!isLoading" class="card-action" style="display: flex; align-items: center; justify-content: space-between">
     <template v-if="user.permissions.modify">
-      <button class="button button--flat" @click="$refs.fileList.createDir()" :aria-label="$t('sidebar.newFolder')"
-        :title="$t('sidebar.newFolder')" style="justify-self: left">
-        <span>{{ $t("sidebar.newFolder") }}</span>
+      <button class="button button--flat" @click="$refs.fileList.createDir()" :aria-label="$t('files.newFolder')"
+        :title="$t('files.newFolder')" style="justify-self: left">
+        <span>{{ $t("files.newFolder") }}</span>
       </button>
     </template>
     <div>
-      <button class="button button--flat button--grey" @click="closeHovers" :aria-label="$t('buttons.cancel')"
-        :title="$t('buttons.cancel')">
-        {{ $t("buttons.cancel") }}
+      <button class="button button--flat button--grey" @click="closeHovers" :aria-label="$t('general.cancel')"
+        :title="$t('general.cancel')">
+        {{ $t("general.cancel") }}
       </button>
       <button :disabled="destContainsSrc" class="button button--flat" @click="performOperation"
-        :aria-label="$t(`buttons.${operation}`)" :title="$t(`buttons.${operation}`)">
-        {{ $t(`buttons.${operation}`) }}
+        :aria-label="operation === 'move' ? $t('general.move') : $t('general.copy')" 
+        :title="operation === 'move' ? $t('general.move') : $t('general.copy')">
+        {{ operation === 'move' ? $t('general.move') : $t('general.copy') }}
       </button>
     </div>
   </div>
@@ -220,9 +221,9 @@ export default {
         ] : undefined
           };
         if (this.operation === "move") {
-          notify.showSuccess(this.$t(`prompts.moveSuccess`), buttonProps);
+          notify.showSuccess(this.$t("prompts.moveSuccess"), buttonProps);
         } else {
-          notify.showSuccess(this.$t(`prompts.copySuccess`), buttonProps);
+          notify.showSuccess(this.$t("prompts.copySuccess"), buttonProps);
         }
       } finally {
         this.isLoading = false; // Hide loading spinner