+++ frontend/src/views/settings/Global.vue
@@ -7,7 +7,7 @@
   <div class="card-content"> {{ $t('settings.emptyGlobal') }} </div>
 
   <div class="card-action">
-    <input class="button button--flat" type="submit" :value="$t('buttons.update')" />
+    <input class="button button--flat" type="submit" :value="$t('general.update')" />
   </div>
 </template>